requirements.txt
fastapi==0.117.1
pydantic==2.11.9
requests==2.32.5
uvicorn==0.36.0
supabase==2.19.0
python-dotenv==1.1.1
Pillow==10.4.0
transformers==4.40.0
torch==2.3.0
sentence-transformers==2.3.1
